Conflicts:
	automation/Common.py
	src/backend/CMakeLists.txt
	src/backend/validations/SATest/CMakeLists.txt

+++ opencl-intel/devices/common/omp2tbb/schedulers.cpp
@@ -8,20 +8,20 @@ extern "C" int omp_get_thread_num();
 
 template < typename T >
 static void static_init( ident_t *loc, kmp_int32 gtid, kmp_int32 schedtype,
-						kmp_int32 *plastiter, T *plower, T *pupper,	T *pstride, T incr, T chunk )
+                        kmp_int32 *plastiter, T *plower, T *pupper,    T *pstride, T incr, T chunk )
 {
-	omp2tbb::global_thread_table::thread_info_entry* entry = omp2tbb::g_thread_info->get_master_entry( gtid );
-	kmp_int32 ltid = omp_get_thread_num();
+    omp2tbb::global_thread_table::thread_info_entry* entry = omp2tbb::g_thread_info->get_master_entry( gtid );
+    kmp_int32 ltid = omp_get_thread_num();
 
-	unsigned int num_threads = entry->my_concurency;
+    int num_threads = (int)entry->my_concurency;
 
     if ( num_threads == 1 ) {
-		// If running on single thread, this is the single and the last iteration
+        // If running on single thread, this is the single and the last iteration
         *plastiter = TRUE;
         return;
     }
 
-	T iter_count;
+    T iter_count;
 
     // Compute number of iterations
     if ( incr == 1 ) {
@@ -36,18 +36,18 @@ static void static_init( ident_t *loc, kmp_int32 gtid, kmp_int32 schedtype,
         }
     }
 
-	switch ( schedtype )
-	{
-	case omp2tbb::kmp_sch_static: {
+    switch ( schedtype )
+    {
+    case omp2tbb::kmp_sch_static: {
             T chunk = iter_count / num_threads;
             T leftover = iter_count % num_threads;
             *plower += incr * ( ltid * chunk + ( ltid < leftover ? ltid : leftover ) );
             *pupper = *plower + chunk * incr - ( ltid < leftover ? 0 : incr );
             *plastiter = ( ltid == num_threads - 1 );
-		}
-		break;
+        }
+        break;
 
-	case omp2tbb::kmp_sch_static_chunked: {
+    case omp2tbb::kmp_sch_static_chunked: {
             if ( chunk < 1 ) {
                 chunk = 1;
             }
@@ -62,18 +62,18 @@ static void static_init( ident_t *loc, kmp_int32 gtid, kmp_int32 schedtype,
     default:
         assert(0 && "__kmpc_for_static_init: unknown scheduling type" );
         break;
-	}
+    }
 }
 
 extern "C" LIBOMP2TBB_API void __kmpc_for_static_init_8( ident_t *loc, kmp_int32 gtid, kmp_int32 schedtype, kmp_int32 *plastiter,
                       kmp_int64 *plower, kmp_int64 *pupper,
                       kmp_int64 *pstride, kmp_int64 incr, kmp_int64 chunk )
 {
-	static_init<kmp_int64>(loc, gtid, schedtype, plastiter, plower, pupper, pstride, incr, chunk);
+    static_init<kmp_int64>(loc, gtid, schedtype, plastiter, plower, pupper, pstride, incr, chunk);
 }
  
 extern "C" LIBOMP2TBB_API void __kmpc_for_static_fini(ident_t* loc, kmp_int32 global_tid)
 {
-	// nothing to do here
-	return;
+    // nothing to do here
+    return;
 }
